ISIS-2177: remove PersistenceSessionServiceInternal
